update of barchart

+++ sound.h
@@ -1,11 +1,12 @@
 // Constants defination
+//#define DEBUG
 #define RATE 16000	//samples per second
 #define CMD "arecord -r16000 -c1 -f S16_LE -d1 -q test.wav"
 // dtata structures
 struct WAVHDR{
 	char ChunkID[4];	//it has to be "RIFF"
 	int ChunkSize;		//4-byte number
-	char Format[4];		// it has to be "WAVE"
+	char Format [4];		// it has to be "WAVE"
 
 	char Subchunk1ID[4];	//"fmt"
 	int Subchunk1Size;	//PCM = 16
@@ -13,7 +14,7 @@ struct WAVHDR{
 	short NumChannels;	// should be 1 for mono
 	int SampleRate;		//16000
 	int ByteRate;		//16000*NumChannels*BitsPerSample/8
-	int BlockAlign;		//NumChannels*BitsPerSample/8
+	short BlockAlign;		//NumChannels*BitsPerSample/8
 	short BitsPerSample;	//in our app, 16 (-f S16_LE)
 
 	char Subchunk2ID[4];	// "data"
